Added ContactPhotoCache
This new class properly encapsulates the ContactPhoto logic.  If the
contact doesn't have a photo, don't keep querying it.  If it does, add
it to the LRU cache.  Also increase the size of the cache to 1/4 instead
of 1/8.  For my ~75 contacts with birthdays, I don't get any more GCs.
Hope that's typical!